Add methods for setting the account after an instance is created
WARNING: this commit breaks API compatibility for the constructor.
The new method setAccount allows to set an account. A new
constructor without the accountName parameter is also available.
We can now create an AppEngineClient instance before the account
is known: the account will be set prior to executing a request.