Add a method for starting the permission dialog + Fix network errors
If the user must grant its permission, an error is thrown. In this error,
an intent is now available to start the permission dialog. Inside an
activity, this intent should be started for asking the permission to
use the authentication token. If the intent is not started (because
we are inside a background service), a notification is still displayed
to start the permission dialog without interrupting the user activity.

Network errors occured because the HTTP responses were not fully
consumed.